refactor(billing): extract Stripe decoding into utils + subscription-labels composable
